feat: adds optional :encoding option to indexers
Defaults to :raw, which writes the index value directly to disk. This could be problematic if the value contains invalid file chars. Otherwise, it can be :hash, or {:hash, hash_alorigthm), which hashs the value and writes the base16 encoded hash value to disk. This prevents invalid file chars, but making the file layout a bit more obscure. The default hash is :sha.