feat: append multiple events
Simply api to append/2 passing the events and opts.

Moved :stream to opts, which be used to enrich each event.
Adds :expectation, which can be :none, or an integer representing the expected position.

Adds the concept of event_refs, a 2-tuple {event_id, event_pos}.

Sends event_refs to EventLedger commit instead of full events.
Sends event_refs to EventIndexManager to index the appended events, instead of directly working with the process group.

Adds read_event/1 to EventReader which reads an event by id. This is used by EventIndexManager to beforing sending to each indexer.